Refactor: Move create metering point handling to separate namespace
