+++ source/Energinet.DataHub.MeteringPoints.Application/Create/Address.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MeteringPoints.Application
+namespace Energinet.DataHub.MeteringPoints.Application.Create
 {
     // TODO: Add all the remaining address fields!
     public record Address(
