+++ source/Energinet.DataHub.MeteringPoints.Application/Validation/CreateMeteringPointRuleSet.cs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.MeteringPoints.Application.Create;
 using Energinet.DataHub.MeteringPoints.Application.Validation.Rules;
 using FluentValidation;
