+++ source/Energinet.DataHub.MeteringPoints.Application/Validation/Rules/MeteringPointSubTypeMustBeValidRule.cs
@@ -15,6 +15,7 @@
 using System;
 using System.Collections.Generic;
 using System.Data;
+using Energinet.DataHub.MeteringPoints.Application.Create;
 using Energinet.DataHub.MeteringPoints.Application.Validation.ValidationErrors;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using FluentValidation;
