+++ source/Energinet.DataHub.MeteringPoints.Application/Validation/Rules/MeteringPointTypeValidRule.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using System.Linq;
+using Energinet.DataHub.MeteringPoints.Application.Create;
 using Energinet.DataHub.MeteringPoints.Application.Validation.ValidationErrors;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
