+++ source/Energinet.DataHub.MeteringPoints.Application/Validation/Rules/PhysicalStateRule.cs
@@ -14,6 +14,7 @@
 
 using System.Collections.Generic;
 using System.Linq;
+using Energinet.DataHub.MeteringPoints.Application.Create;
 using Energinet.DataHub.MeteringPoints.Application.Validation.ValidationErrors;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
