+++ source/Energinet.DataHub.MeteringPoints.Application/Validation/Rules/PowerLimitRule.cs
@@ -14,6 +14,7 @@
 
 using System;
 using System.Linq;
+using Energinet.DataHub.MeteringPoints.Application.Create;
 using Energinet.DataHub.MeteringPoints.Application.Validation.ValidationErrors;
 using FluentValidation;
