+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/TestHost.cs
@@ -46,7 +46,7 @@ using SimpleInjector.Lifestyles;
 using Xunit;
 using Xunit.Categories;
 using ConnectMeteringPoint = Energinet.DataHub.MeteringPoints.Application.Connect.ConnectMeteringPoint;
-using CreateMeteringPoint = Energinet.DataHub.MeteringPoints.Application.CreateMeteringPoint;
+using CreateMeteringPoint = Energinet.DataHub.MeteringPoints.Application.Create.CreateMeteringPoint;
 
 namespace Energinet.DataHub.MeteringPoints.IntegrationTests
 {
